requirements.txt
ffmpeg-python==0.2.0
ffprobe==0.5
mediapipe==0.9.2.1
opencv-python==4.7.0.72
opencv-contrib-python==4.7.0.72